Move some functionality from server index to helpers.js
